Simplifies canceling workflow
Instead of multiple workflow-runs generated for different workflows
this change replces it with single workflow, cancelling all others.
It is triggered by only one CI-IT workflow, because this one is
always triggered when all others are. It also sets parallelism
of the matrix to 4, so that there are not too many jobs in
parallel taken when this workflow runs.