The Cancel step is now optimised in one job
The cancel steps were tried in separate jobs but it causes
an unnecessary increase in the number of jobs blocking the queue
even if we limited parallelism. This PR combines all the cancel
actions into a single job with multiple steps.

While this is a bit more copy&pasting and less DRY it is also
far more optimised and "aggressive" - i.e. once any cancel
job starts to be running it will immediately cancel all duplicates
for all workflows.