[Enhancement] Bump grpc and protobuff version to allow building on Mac M1
Upgrade to the latest grpc-java 1.42.1 which is only compatible and [tested against](https://github.com/grpc/grpc-java/pull/8551#discussion_r714182190) 3.17.2, but we can assume that 3.17.3 (which has osx-aarch_64 compatible maven artifacts) is save to use.

We could have also upgrade to the latest grpc-java 1.42.1 and latest protoc 3.19.1, but it can not be directly done, it needs to tweak the grpcs-java internal dependencies. So we choose the safest option and discarded this option because of the following reasons:
 - Protoc 3.19.1 generates code that relies on methods not available in protobuf-java 3.17.2 ([“issue”](https://github.com/protocolbuffers/protobuf/issues/9236#issue-1059594027)) (which java-grpc relies on). So we need to tell grpc-java via dependency management to use 3.19.1 instead of 3.17.2.

Signed-off-by: Marc Navarro Sonnenfeld <marcnavarro@tetrate.io>