restrict github actions from running on all PRs
Modified GitHub Actions workflow to only trigger the CodeQL analysis job when changes are made to Java files. This ensures that the analysis focuses specifically on Java code. Enhanced the Dead Link Checker GitHub Actions workflow to only run when changes are made in the 'docs' directory. This allows us to focus the link checking specifically on the documentation files.